+++ includes/classes/SupportMonitor/Monitor.php
@@ -354,7 +354,8 @@ class Monitor extends Singleton {
 			'data'       => $data,
 			'type'       => sanitize_text_field( $type ),
 			'group'      => sanitize_text_field( $group ),
-			'message_id' => md5( $setting['api_key'] . home_url() . microtime( true ) ),
+			'message_id' => md5( $setting['api_key'] . microtime( true ) . wp_rand( 0, 10000 ) ),
+
 		];
 
 		return apply_filters( 'tenup_support_monitor_message', $message );
@@ -579,8 +580,13 @@ class Monitor extends Singleton {
 	 * @return string
 	 */
 	public function get_wp_version() {
-		$data = get_preferred_from_update_core();
-		return $data->version;
+		global $wp_version;
+
+		if ( ! empty( $wp_version ) ) {
+			return $wp_version;
+		}
+
+		return null;
 	}
 
 	/**